Generate the visibility data from a given Chunk.
The visibility data is what will(TODO) get send to the Client for further usage.
On the client the visiblity data will(TODO) will be used to generate the chunk mesh and lighting.
The visiblity data contains all blocks that are visible from a given chunk and also data about which of its neighbors blocks the block(to later cull faces that are below the terrain).